modify nip3041 doc + reorganized the client-cli
